backend/requirements.txt
# Core framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic-settings==2.7.0
email-validator==2.2.0

# Gemini
google-generativeai==0.8.3

# Video metadata extraction
mutagen==1.47.0
hachoir==3.3.0

# Image processing — TradePhotoAnalyzer (resize, sharpness detection, JPEG re-encoding)
Pillow==11.1.0

# Supabase
supabase==2.10.0

# HTTP client (Smarty API, future integrations)
httpx==0.27.2

# [SECURITY: code-review] Rate limiting middleware to protect auth endpoints from
# brute-force and email-bombing attacks.
slowapi==0.1.9

# Stripe payment processing (escrow layer)
stripe>=7.0.0

# Web Push notifications (VAPID / RFC 8030)
pywebpush>=2.0.0
